requirements.txt
Django==2.0
django-debug-toolbar==1.10.1
django-filebrowser==3.10.1
django-grappelli==2.11.1
Pillow==5.3.0

